backend/requirements/requirements-dev.txt
aiofiles==23.2.1
alembic==1.12.1
annotated-types==0.6.0
anyio==3.7.1
async-timeout==4.0.3
asyncpg==0.29.0
black==23.11.0
certifi==2023.7.22
click==8.1.7
exceptiongroup==1.1.3
factory-boy==3.3.0
Faker==20.0.3
fastapi==0.104.1
greenlet==3.0.1
h11==0.14.0
httpcore==1.0.2
httpx==0.25.1
idna==3.4
iniconfig==2.0.0
Mako==1.3.0
MarkupSafe==2.1.3
mypy-extensions==1.0.0
packaging==23.2
pathspec==0.11.2
platformdirs==4.0.0
pluggy==1.3.0
postgres==4.0
psycopg2-binary==2.9.9
psycopg2-pool==1.1
pydantic==2.4.2
pydantic_core==2.10.1
pytest==7.4.3
pytest-asyncio==0.21.1
python-dateutil==2.8.2
python-dotenv==1.0.0
python-multipart==0.0.6
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.23
starlette==0.27.0
tomli==2.0.1
typing_extensions==4.8.0
uvicorn==0.24.0.post1
